+++ AutoFTG/release_notes.md
@@ -12,4 +12,20 @@ Release notes for AutoFTG python scripts
 ## v1.1
 
 
+---
+
+## v1.7.5
+
+- Added options to change default settings for:
+  - Change default project folder
+  - Change default camera calibration
+  - Change default point filtering spacing
+  - Change default point sampling
+
+---
+
+## v 1.7.6
+
+- Reorder manus for default settings
+